Do not cancel all jobs in a workflow if one fails.
GitHub will usually cancel all running jobs in a matrix as soon
as the first job in that matrix fails. However, this may be
counter-productive, because even jobs that may pass will get
cancelled this way, making it harder to figure out which jobs
actually fail and which jobs would pass. For more information on
the `fail-fast` configuration option see
<https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstrategyfail-fast>.